(build): the linter rides inside javac, and a proven bug pattern fails the build (#102)
Error Prone, pinned in the parent pom, its findings promoted to compile
errors and the generated codecs exempt, so linting is part of compiling
rather than a separate opinion. The first sweep's forty one findings:
fixed where the advice was right (explicit precedence in the tie-breaks,
split limits, asSubclass over a blind cast, an overflow made impossible,
charsets and locales said out loud, a dead constant removed, NavigableMap
where that is all the callers ask), and refused at the site with the
reason beside it where the finding was the design (the lean twin's
unread level fields are the layout, and the decode-only engine counts
by ordinal on purpose).

+++ java/matching-conformance/src/main/java/io/github/giovanicaprison/matching/conformance/CommandWriter.java
@@ -135,7 +135,7 @@ final class CommandWriter {
     cancel.frame().instrumentId(INSTRUMENT_ID).sequence(sequence);
     cancel
         .clientOrderId(reference)
-        .participantId((int) number(options, "p", references.participant(reference)));
+        .participantId(number(options, "p", references.participant(reference)));
     return length(cancel.encodedLength());
   }
 
@@ -155,7 +155,7 @@ final class CommandWriter {
     final Map<String, String> options = options(arguments);
     massCancel.wrapAndApplyHeader(buffer, 0, header);
     massCancel.frame().instrumentId(INSTRUMENT_ID).sequence(sequence);
-    massCancel.clientOrderId(0).participantId((int) number(options, "p"));
+    massCancel.clientOrderId(0).participantId(number(options, "p"));
     return length(massCancel.encodedLength());
   }
